// angular-frontend/src/app/clients/open-f1/models/meeting/meeting.interface.ts
export interface IMeeting {
    circuitKey: number
    circuitShortName: string
    countryCode: string
    countryKey: number
    countryName: string
    dateStart: Date
    gmtOffset: number
    location: string
    meetingCode: string
    meetingKey: number
    meetingName: string
    meetingOfficialName: string
    year: number
}